change coordinate date type to dbl and update seed values

+++ TrailsClient/Models/TrailMarker.cs
@@ -12,9 +12,9 @@ namespace TrailsClient.Models
 
     public string Name { get; set; }
 
-    public int Longitude { get; set; }
+    public double Longitude { get; set; }
 
-    public int Latitude { get; set; }
+    public double Latitude { get; set; }
 
     public bool isTrailhead { get; set; }
